% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjx0y0.R
\name{adjx0y0}
\alias{adjx0y0}
\title{Calculate Origin of Axis based on Origin of Subplot}
\usage{
adjx0y0(p, xory, secondary)
}
\arguments{
\item{p}{The PSTricks object.}

\item{xory}{A character 'x' or 'y' designating the axis.}

\item{secondary}{A flag to designate a secondary axis.}
}
\value{
An origin.
}
\description{
Calculate Origin of Axis based on Origin of Subplot
}
