% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{aes}
\alias{aes}
\title{Construct Aesthetic Mappings}
\usage{
aes(...)
}
\arguments{
\item{...}{Comma separated mappings such as in the example below.}
}
\value{
A structure containing the mapping.
}
\description{
Construct Aesthetic Mappings
}
\details{
Note: \code{aes()} does not evaluate right hand sides of mappings.
}
\examples{
aes("x=time")
}
\seealso{
\code{\link[=geom_set]{geom_set()}} for an example.
}
