% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{circlenode}
\alias{circlenode}
\title{Put Stuff in a Circle}
\usage{
circlenode(p = NULL, name, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{name}{The name of the node.}

\item{stuff}{Stuff to put in a box at the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Circle
}
\seealso{
\code{\link[=ovalnode]{ovalnode()}} for an example.
}
