% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{clipbox}
\alias{clipbox}
\title{Put Stuff in a Box with Clipping}
\usage{
clipbox(p = NULL, stuff, dim = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The stuff to put in the box.}

\item{dim}{Distance between the box and clipping.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Box with Clipping
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,clipbox(,"\\\\parbox[t][1cm][t]{2cm}{One of the best
        new plays I have seen all all year}",-0.1))
}
