% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{cnode}
\alias{cnode}
\title{Create Circle Node}
\usage{
cnode(p = NULL, x = NULL, y = NULL, radius, name, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the node.}

\item{radius}{Radius of the circle.}

\item{name}{The name of the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Create Circle Node
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    cnode(0,1,0.25,"A") \%>\%
    pnode(3,0,"B") \%>\%
    ncline("A","B",arrows="<-")
}
