% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{cnodeput}
\alias{cnodeput}
\title{Put Stuff in a Circle}
\usage{
cnodeput(
  p = NULL,
  x = NULL,
  y = NULL,
  name,
  stuff,
  par = NULL,
  angle = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the node.}

\item{name}{The name of the node.}

\item{stuff}{Stuff to put in a box at the node.}

\item{par}{PSTricks parameter string.}

\item{angle}{Angle to put the stuff with.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Circle
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    cnodeput(,,"A","X",angle=45)
}
