% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{cput}
\alias{cput}
\title{Put Stuff in a Circle}
\usage{
cput(p = NULL, x, y, stuff, par = NULL, angle = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the center of the circle.}

\item{stuff}{The stuff to put in the box.}

\item{par}{PSTricks parameter string.}

\item{angle}{Rotation to apply to the stuff.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Circle
}
\examples{
pppicture(PSTricks(),2,1,par="showgrid=true") \%>\%
    cput(1,.5,"\\\\large $K_1$","doubleline=true")
}
