% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{Rnode}
\alias{Rnode}
\title{Put Stuff in a Box at a Node}
\usage{
Rnode(p = NULL, name, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{name}{The name of the node.}

\item{stuff}{Stuff to put in a box at the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Box at a Node
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
   rput(8,4,paste("\\\\Large",Rnode(,"A","sp"),"\\\\hskip 2cm",Rnode(,"B","Bit"))) \%>\%
   ncline("A","B")
}
