% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cx}
\alias{cx}
\title{Convert Unscaled \code{x} Values to Scaled}
\usage{
cx(p, x, logx = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x}{Unscaled data.}

\item{logx}{Flag to request log(10) transformation.}
}
\value{
Scaled data.
}
\description{
Convert Unscaled \code{x} Values to Scaled
}
