% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cy}
\alias{cy}
\title{Convert Unscaled \code{y} Values to Scaled}
\usage{
cy(p, y, logy = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{y}{Unscaled data.}

\item{logy}{Flag to request log(10) transformation.}
}
\value{
Scaled data.
}
\description{
Convert Unscaled \code{y} Values to Scaled
}
