% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{degrees}
\alias{degrees}
\title{Set Unit for Angles}
\usage{
degrees(p, degrees = 360)
}
\arguments{
\item{p}{The PSTricks object.}

\item{degrees}{The number of units in a circle.}
}
\value{
The updated PSTricks object.
}
\description{
Set Unit for Angles
}
\seealso{
\code{\link[=ppsetpolar]{ppsetpolar()}}.
}
