% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{dotnode}
\alias{dotnode}
\title{Create a Dot Node}
\usage{
dotnode(p = NULL, x = NULL, y = NULL, name, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the node.}

\item{name}{The name of the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Create a Dot Node
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    dotnode(,,"A","dotstyle=triangle*,dotscale=2 1") \%>\%
    dotnode(3,2,"B","dotstyle=+") \%>\%
    ncline("A","B","nodesep=3pt")
}
