% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2e.R
\name{endP2E}
\alias{endP2E}
\title{End PSTtoEPS Feature}
\usage{
endP2E(p, fileplot = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{fileplot}{Flag to indicate \code{cat}ed values will be used for \code{fileplot}.}
}
\value{
The updated PSTricks object.
}
\description{
End PSTtoEPS Feature
}
