% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picture.R
\name{endpppicture}
\alias{endpppicture}
\title{Close the Picture}
\usage{
endpppicture(p, ending = "")
}
\arguments{
\item{p}{The PSTricks object.}

\item{ending}{String to end the pppicture environment with.}
}
\value{
The updated PSTricks object.
}
\description{
Close the Picture
}
