% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{endpspicture}
\alias{endpspicture}
\title{End Picture Environment}
\usage{
endpspicture(p = NULL)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
End Picture Environment
}
\seealso{
\code{\link[=pspicture]{pspicture()}} for an example.
}
