% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{everypsbox}
\alias{everypsbox}
\title{Prepend String to every \code{psbox}}
\usage{
everypsbox(p, s)
}
\arguments{
\item{p}{The PSTricks object.}

\item{s}{The string to prepend.}
}
\value{
The updated PSTricks object.
}
\description{
Prepend String to every \code{psbox}
}
\examples{
everypsbox(PSTricks(), "\\\\Large")$lines[[1]]
}
