% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{fnode}
\alias{fnode}
\title{Create a Frame Node}
\usage{
fnode(p = NULL, x = NULL, y = NULL, name, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Optional coordinates of the center.}

\item{name}{The name of the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Create a Frame Node
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    fnode(,,"A") \%>\%
    fnode(2,2,"B","framesize=1 5pt",TRUE) \%>\%
    ncline("A","B","nodesep=3pt")
}
