% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_abline}
\alias{geom_abline}
\title{Draw Straight Line}
\usage{
geom_abline(p, slope = 1, intercept = 0, par = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{slope}{The slope of the line, or an \code{lm} object.}

\item{intercept}{The intercept of the line.}

\item{par}{PSTricks parameter string.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Straight Line
}
\examples{
PSTricks() \%>\%
    pppicture(16,9) \%>\%
    ppsetlogxy() \%>\%
    geom_dots(aes(x=hp,y=mpg),mtcars,par="dotstyle=Bo") \%>\%
    geom_abline(lm(log10(mpg)~log10(hp),data=mtcars),par="linecolor=red") \%>\%
    geom_hline(20,par="linecolor=green") \%>\%
    geom_vline(100,par="linecolor=blue")
# Note that log10 needs to be used for lm with log axes
}
