% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_ccurve}
\alias{geom_ccurve}
\title{Connect Observations using Smooth Lines}
\usage{
geom_ccurve(
  p,
  mapping = NULL,
  data = NULL,
  par = NULL,
  dodge = 0,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Connect Observations using Smooth Lines
}
\examples{
geom_ccurve(PSTricks(),
    data=data.frame(x=c(.5,3.5,3.5,.5),y=c(0,1,0,1)),
        par="showpoints=true") \%>\%
    xlim(0,4) \%>\% ylim(-0.5,1.5) \%>\%
    geom_grid()
}
\seealso{
\code{\link[=psccurve]{psccurve()}} for the base version.
}
