% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_curve}
\alias{geom_curve}
\title{Connect Observations using Smooth Lines}
\usage{
geom_curve(p, mapping = NULL, data = NULL, par = NULL, dodge = 0, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Connect Observations using Smooth Lines
}
\examples{
PSTricks() \%>\%
    newrgbcolor("verylightgray",.9,.9,.9) \%>\%
    ppsetmargins(mrgaxes=0) \%>\%
    geom_grid("linestyle=dotted,linecolor=gray",
        background="verylightgray") \%>\%
    geom_curve(data=data.frame(x=c(0,.7,3.3,4,.4),y=c(1.3,1.8,.5,1.6,.4)),
        par="showpoints=true") \%>\%
    geom_legend("top right","showpoints=true") \%>\%
    xlim(-1,5) \%>\% ylim(0,2)
# Note that autoscaling which uses the data only does not work optimally
}
\seealso{
\code{\link[=pscurve]{pscurve()}} for the base version.
}
