% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_dots}
\alias{geom_dots}
\title{Plot Dots}
\usage{
geom_dots(p, mapping = NULL, data = NULL, par = NULL, dodge = 0, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Plot Dots
}
\examples{
geom_dots(PSTricks(),data=data.frame(x=c(0,1,2),y=c(1,1,1),
    par=paste0("dotstyle=",c('*','o','Bo'))))
}
\seealso{
\code{\link[=psdots]{psdots()}} for the base version and \code{\link[=geom_abline]{geom_abline()}} for another example.
}
