% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_ecurve}
\alias{geom_ecurve}
\title{Connect Observations using Smooth Lines}
\usage{
geom_ecurve(
  p,
  mapping = NULL,
  data = NULL,
  par = NULL,
  dodge = 0,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Connect Observations using Smooth Lines
}
\examples{
PSTricks() \%>\%
    pppicture(16,9,star=TRUE) \%>\%
    geom_ecurve(data=data.frame(x=c(.125,.25,.5,1,2,4,8),
                                y=c(8,4,2,1,.5,.25,.125)),
                par="showpoints=true") \%>\%
    xlim(0,4) \%>\% ylim(0,4) \%>\%
    geom_grid()
}
\seealso{
\code{\link[=psecurve]{psecurve()}} for the base version.
}
