% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_errorbar}
\alias{geom_errorbar}
\title{Vertical Errorbars}
\usage{
geom_errorbar(
  p,
  mapping = NULL,
  data = NULL,
  par = NULL,
  width = 0.1,
  dodge = 0
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x}, \code{y}, \code{ymin}, and \code{ymax}.}

\item{data}{Data frame with values for the error bars.}

\item{par}{PSTricks parameters.}

\item{width}{Horizontal width of the error bars.}

\item{dodge}{Horizontal offset.}
}
\value{
The updated PSTricks object.
}
\description{
Vertical Errorbars
}
\examples{
pppicture(PSTricks(),16,9,data=data.frame(x=c(1,2,3,4),y=c(2,4,6,8),
                                          ymin=c(1,2,3,4),ymax=c(3,6,9,12))) \%>\%
    geom_set("linecolor=blue") \%>\%
    geom_line(par="showpoints=true",dodge=-0.125) \%>\%
    geom_errorbar(dodge=-0.125) \%>\%
    geom_set("linecolor=green") \%>\%
    geom_line(par="showpoints=true",dodge=0.125) \%>\%
    geom_errorbar(aes(ymin=NA),dodge=0.125)
}
