% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_everypsbox}
\alias{geom_everypsbox}
\title{Set \code{everypsbox} during Geom Processing}
\usage{
geom_everypsbox(p, par = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{par}{Stuff to apply to a \code{psbox}.}
}
\value{
The updated PSTricks object.
}
\description{
Set \code{everypsbox} during Geom Processing
}
\seealso{
\code{\link[=everypsbox]{everypsbox()}} for the base version and \code{\link[=geom_set]{geom_set()}} for an example.
}
