% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_frame}
\alias{geom_frame}
\title{Draw Frames}
\usage{
geom_frame(p, mapping = NULL, data = NULL, par = NULL, dodge = 0, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Frames
}
\examples{
geom_frame(PSTricks(),
    data=data.frame(x0=c(0,1),x1=c(4,.5),y0=c(0,.5),y1=c(2,1.5),
        par=c("linewidth=2pt,framearc=.3,fillstyle=solid,fillcolor=lightgray",
           "linecolor=white"),
        star=c(FALSE,TRUE)))
}
\seealso{
\code{\link[=psframe]{psframe()}} for the base version.
}
