% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_framebox}
\alias{geom_framebox}
\title{Add Frameboxes}
\usage{
geom_framebox(
  p,
  mapping = NULL,
  data = NULL,
  par = NULL,
  refpoint = NULL,
  rotation = NULL,
  dodge = 0,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{refpoint}{The reference point for the stuff.}

\item{rotation}{Rotation to apply to the stuff.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Add Frameboxes
}
\seealso{
\code{\link[=psframebox]{psframebox()}} and \code{\link[=rput]{rput()}} for the base versions and \code{\link[=geom_set]{geom_set()}} for an example.
}
