% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_grid}
\alias{geom_grid}
\title{Draw Grid Lines}
\usage{
geom_grid(p, par = "linestyle=dotted", background = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{par}{PSTricks parameters.}

\item{background}{The background color.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Grid Lines
}
\examples{
geom_grid(PSTricks())
}
\seealso{
\code{\link[=ppgrid]{ppgrid()}} for the base version and \code{\link[=geom_curve]{geom_curve()}} for an example.
}
