% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_hist}
\alias{geom_hist}
\title{Plot a Histogram}
\usage{
geom_hist(
  p,
  mapping = aes(x = breaks, y = counts),
  data = NULL,
  par = "fillcolor=lightgray,fillstyle=solid",
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Either \code{aes(x=breaks,y=counts)} or \code{aes(x=breaks,y=density)}.}

\item{data}{Output of R's \code{hist(..., plot=FALSE)} function.}

\item{par}{PSTricks parameters.}

\item{star}{Flag to use \code{star} version of \code{psframe}.}
}
\value{
The updated PSTricks object.
}
\description{
Plot a Histogram
}
\details{
Issue: The default mapping containing "breaks" and "counts" leads to a NOTE when running "R CMD check".
}
\examples{
geom_hist(PSTricks(),data=hist(mtcars$mpg,plot=FALSE),
    par="fillcolor=cyan,fillstyle=solid")
}
