% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_hline}
\alias{geom_hline}
\title{Draw Horizontal Line}
\usage{
geom_hline(p, yintercept = 0, par = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{yintercept}{The y-intercept of the line.}

\item{par}{PSTricks parameter string.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Horizontal Line
}
\seealso{
\code{\link[=geom_abline]{geom_abline()}} for an example.
}
