% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_legend}
\alias{geom_legend}
\title{Add Legend to Plot}
\usage{
geom_legend(
  p,
  s,
  par = NULL,
  position = "tr",
  dx = 0,
  dy = 0,
  w = 1,
  labelsep = "10pt"
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{s}{The legend text.}

\item{par}{PSTricks parameter string.}

\item{position}{Position for the legend (may be NULL).}

\item{dx, dy}{x and y offsets w.r.t. default position.}

\item{w}{Width of the \code{psline} that belongs to the legend text.}

\item{labelsep}{The distance between the line and the label.}
}
\value{
The updated PSTricks object.
}
\description{
Add Legend to Plot
}
\seealso{
\code{\link[=pplegend]{pplegend()}} for the base version and \code{\link[=geom_curve]{geom_curve()}} for an example.
}
