% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_linewidth}
\alias{geom_linewidth}
\title{Set PSTricks' \code{linewidth} Parameter during Geom Processing}
\usage{
geom_linewidth(p, linewidth = 0.8 * 2.54/72)
}
\arguments{
\item{p}{The PSTricks object.}

\item{linewidth}{The linewidth to use (default the PSTricks default (0.8 pt)).}
}
\value{
The updated PSTricks object.
}
\description{
Set PSTricks' \code{linewidth} Parameter during Geom Processing
}
\seealso{
\code{\link[=pplinewidth]{pplinewidth()}} for the base version and \code{\link[=geom_set]{geom_set()}} for an example.
}
