% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_polygon}
\alias{geom_polygon}
\title{Draw Polygons}
\usage{
geom_polygon(
  p,
  mapping = NULL,
  data = NULL,
  par = NULL,
  dodge = 0,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{par}{PSTricks parameter string.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Polygons
}
\examples{
PSTricks() \%>\%
    geom_polygon(data=data.frame(x=c(0,0,1),y=c(0,2,2)),par="linewidth=1.5pt") \%>\%
    geom_polygon(data=data.frame(x=c(1,1,4,4),y=c(0,2,0,2)),par="linearc=.2",star=TRUE)
# Note that the first coordinate (0,0) for the first polygon has to be given explicitly
}
\seealso{
\code{\link[=pspolygon]{pspolygon()}} for the base version.
}
