% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_rput}
\alias{geom_rput}
\title{Add Text Items}
\usage{
geom_rput(
  p,
  mapping = NULL,
  data = NULL,
  refpoint = NULL,
  rotation = NULL,
  dodge = 0,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{refpoint}{The reference point for the stuff.}

\item{rotation}{Rotation to apply to the stuff.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version (but see \code{geom_framebox()}).}
}
\value{
The updated PSTricks object.
}
\description{
Add Text Items
}
\examples{
geom_rput(PSTricks(),
    aes(x=wt,y=mpg,stuff=stuff),
    cbind(mtcars,stuff=row.names(mtcars)),
    rotation=45,
    star=TRUE)
}
\seealso{
\code{\link[=rput]{rput()}} for the base version.
}
