% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_set}
\alias{geom_set}
\title{Set PSTricks Parameter(s) during Geom Processing}
\usage{
geom_set(p, par)
}
\arguments{
\item{p}{The PSTricks object.}

\item{par}{PSTricks (comma separated) parameter(s).}
}
\value{
The updated PSTricks object.
}
\description{
Set PSTricks Parameter(s) during Geom Processing
}
\examples{
mtcars<-cbind(mtcars,stuff=row.names(mtcars));
PSTricks() \%>\%
    pppicture(16,26) \%>\%
# the following three commands affect the axes
    psset("arrows=c-c") \%>\%
    pplinewidth(.3) \%>\%
    everypsbox("\\\\large") \%>\%
# the following three commands affect the frameboxes
    geom_set("framearc=.3,fillstyle=solid,fillcolor=darkgray") \%>\%
    geom_linewidth(.1) \%>\%
    geom_everypsbox("\\\\green") \%>\%
    geom_framebox(aes(x=wt,y=mpg),mtcars[mtcars$cyl==4,]) \%>\%
    geom_linewidth(.3) \%>\%
    geom_everypsbox("\\\\cyan") \%>\%
    geom_framebox(aes(x=wt,y=mpg),mtcars[mtcars$cyl==6,]) \%>\%
    geom_linewidth(.5) \%>\%
    geom_everypsbox("\\\\red") \%>\%
    geom_framebox(aes(x=wt,y=mpg),mtcars[mtcars$cyl==8,]) \%>\%
    lims(c(1,6),c(10,35)) \%>\%
    labs("Weight (lb/1000)","Fuel efficiency (miles/gallon)") \%>\%
    pplegend("4 cylinders",par="linecolor=green",dx=-3) \%>\%
    pplegend("6 cylinders",par="linecolor=cyan",dx=-3,dy=-.5) \%>\%
    pplegend("8 cylinders",par="linecolor=red",dx=-3,dy=-1)
}
\seealso{
\code{\link[=psset]{psset()}} for the base version.
}
