% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_uput}
\alias{geom_uput}
\title{Add Text Items}
\usage{
geom_uput(
  p,
  mapping = NULL,
  data = NULL,
  refangle = NULL,
  rotation = NULL,
  labelsep = NULL,
  dodge = 0,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{mapping}{Aesthetic mapping from column names to \code{x} and \code{y}.}

\item{data}{Data frame with coordinates of the observations.}

\item{refangle}{The reference angle.}

\item{rotation}{Rotation to apply to the stuff.}

\item{labelsep}{Distance between coordinates and the stuff.}

\item{dodge}{Horizontal offset.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Add Text Items
}
\examples{
geom_uput(PSTricks(),
    aes(x=wt,y=mpg,stuff=stuff),
    cbind(mtcars,stuff=row.names(mtcars)),
    refangle=0,
    rotation=45,
    star=TRUE)
}
\seealso{
\code{\link[=uput]{uput()}} for the base version.
}
