% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_vline}
\alias{geom_vline}
\title{Draw Vertical Line}
\usage{
geom_vline(p, xintercept = 0, par = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{xintercept}{The x-intercept of the line.}

\item{par}{PSTricks parameter string.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Vertical Line
}
\seealso{
\code{\link[=geom_abline]{geom_abline()}} for an example.
}
