% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{icx}
\alias{icx}
\title{Convert Scaled \code{x} Values to Unscaled}
\usage{
icx(p, x, logx = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x}{Scaled data.}

\item{logx}{Flag to request log(10) transformation.}
}
\value{
Unscaled data.
}
\description{
Convert Scaled \code{x} Values to Unscaled
}
