% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{icy}
\alias{icy}
\title{Convert Scaled \code{y} Values to Unscaled}
\usage{
icy(p, y, logy = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{y}{Scaled data.}

\item{logy}{Flag to request log(10) transformation.}
}
\value{
Unscaled data.
}
\description{
Convert Scaled \code{y} Values to Unscaled
}
