% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{labs}
\alias{labs}
\title{Set Axis Labels and Title}
\usage{
labs(p, x, y, title = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{x and y axis labels.}

\item{title}{The title for the plot.}
}
\value{
The updated PSTricks object.
}
\description{
Set Axis Labels and Title
}
\seealso{
\code{\link[=geom_set]{geom_set()}} for an example.
}
