% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{lims}
\alias{lims}
\title{Set x and y Axes Limits}
\usage{
lims(p, x = NULL, y = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{x and y lower and upper axis limits (two-element lists or NULL for automatic).}
}
\value{
The updated PSTricks object.
}
\description{
Set x and y Axes Limits
}
\seealso{
\code{\link[=geom_set]{geom_set()}} for an example.
}
