% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{MakeShortNab}
\alias{MakeShortNab}
\title{Define Short Form Characters}
\usage{
MakeShortNab(p = NULL, char1, char2)
}
\arguments{
\item{p}{The PSTricks object.}

\item{char1}{Short form character for \code{naput}.}

\item{char2}{Short form character for \code{nbput}.}
}
\value{
The updated PSTricks object.
}
\description{
Define Short Form Characters
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    cnode(0,4,".5cm","root") \%>\%
    cnode(3,5.5,"4pt","A",star=TRUE) \%>\%
    cnode(3,2.5,"4pt","C",star=TRUE) \%>\%
    psset("nodesep=3pt,shortput=nab") \%>\%
    MakeShortNab("+","-") \%>\%
    ppappend(paste0(ncline(,"root","A"),"+{$x$}")) \%>\%
    ppappend(paste0(ncline(,"root","C"),"-{$y$}"))
# so short forms are not elegantly implemented
}
