% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{MakeShortTablr}
\alias{MakeShortTablr}
\title{Define Short Form Characters}
\usage{
MakeShortTablr(p = NULL, char1, char2, char3, char4)
}
\arguments{
\item{p}{The PSTricks object.}

\item{char1}{Short form character for \code{taput}.}

\item{char2}{Short form character for \code{tbput}.}

\item{char3}{Short form character for \code{tlput}.}

\item{char4}{Short form character for \code{trput}.}
}
\value{
The updated PSTricks object.
}
\description{
Define Short Form Characters
}
\seealso{
See \code{\link[=MakeShortNab]{MakeShortNab()}} for how to use short forms.
}
