% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.list}
\alias{merge.list}
\title{Merge Two Lists}
\usage{
\method{merge}{list}(x, y, ...)
}
\arguments{
\item{x}{The first list.}

\item{y}{The second list, used to add missing elements in the first list.}

\item{...}{Not used.}
}
\value{
The merged lists.
}
\description{
Merge Two Lists
}
\examples{
merge(list(a=3,b=4),list(a=30,c=40))
}
