% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{multirput}
\alias{multirput}
\title{Put Copies of Stuff}
\usage{
multirput(
  p = NULL,
  x,
  y,
  n,
  stuff,
  angle = NULL,
  refpoint = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the stuff.}

\item{n}{Number of copies.}

\item{stuff}{Stuff to put at the reference point.}

\item{angle}{Angle for the copies.}

\item{refpoint}{The reference point for the stuff.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Copies of Stuff
}
\examples{
pppicture(PSTricks(),3,3) \%>\%
    multirput(c(.5,.3),c(0,.1),12,'*')
}
