% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ncarc}
\alias{ncarc}
\title{Draw an Arc Between Two Nodes}
\usage{
ncarc(p = NULL, nodeA, nodeB, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw an Arc Between Two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    cnodeput(0,0,"A","X") \%>\%
    cnodeput(3,2,"B","Y") \%>\%
    psset("nodesep=3pt") \%>\%
    ncarc("A","B",arrows="->") \%>\%
    ncarc("B","A",arrows="->")
}
