% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ncarcbox}
\alias{ncarcbox}
\title{Enclose Two Nodes in Curved Box}
\usage{
ncarcbox(p = NULL, nodeA, nodeB, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Enclose Two Nodes in Curved Box
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(.5,0,rnode(,"A","1"),"bl") \%>\%
    rput(3.5,2,rnode(,"B","2"),"tr") \%>\%
    ncarcbox("A","B","nodesep=.2cm,boxsize=.4,linearc=.4,arcangle=50")
}
