% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{nccircle}
\alias{nccircle}
\title{Draw a Circle between a Node and Itself}
\usage{
nccircle(p = NULL, node, radius, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{node}{Name of the node.}

\item{radius}{Radius of the circle.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Circle between a Node and Itself
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rnode("A","\\\\textbf{back}") \%>\%
    nccircle("A",".7cm","nodesep=3pt","->")
}
