% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-coil.R
\name{nccoil}
\alias{nccoil}
\title{Draw a Coil between two Nodes}
\usage{
nccoil(p = NULL, nodeA, nodeB, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Coil between two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-coil"),c(-1,5),c(-1,4),par="showgrid=true") \%>\%
    cnode(.5,.5,.5,"A") \%>\%
    cnode(3.5,2.5,.5,"B","fillstyle=solid,fillcolor=lightgray") \%>\%
    nccoil("A","B","coilwidth=.3","<->")
# Note that the `pst-node` macro package does not have to be specified.
}
