% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{nccurve}
\alias{nccurve}
\title{Draw a Bezier Curve between Two Nodes}
\usage{
nccurve(p = NULL, nodeA, nodeB, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Bezier Curve between Two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(0,0,rnode(,"A",psframebox(,"Node A")),"bl") \%>\%
    rput(4,3,ovalnode(,"B","Node B"),"tr") \%>\%
    nccurve("A","B","angleB=180")
}
