% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ncdiagg}
\alias{ncdiagg}
\title{Draw Line Segments Between Two Nodes}
\usage{
ncdiagg(p = NULL, nodeA, nodeB, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Line Segments Between Two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    cnode(0,4,"12pt","a") \%>\%
    rput(3,5,rnode(,"b","H"),"l") \%>\%
    rput(3,3,rnode(,"c","T"),"l") \%>\%
    ncdiagg("b","a","angleA=180,armA=1.5,nodesepA=3pt") \%>\%
    ncdiag("c","a","angleA=180,armA=1.5,armB=0,nodesepA=3pt")
}
