% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ncline}
\alias{ncline}
\title{Draw a Line Between Two Nodes}
\usage{
ncline(p = NULL, nodeA, nodeB, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Line Between Two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(0,0,rnode(,"A","Idea 1"),"bl") \%>\%
    rput(4,3,rnode(,"B","Idea 2"),"tr") \%>\%
    ncline("A","B","nodesep=3pt","<->")
}
