% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ncloop}
\alias{ncloop}
\title{Draw Line Segments Between a Node and Itself}
\usage{
ncloop(p = NULL, nodeA, nodeB, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nodeA, nodeB}{Names of the node.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Line Segments Between a Node and Itself
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rnode("a",psframebox(,"\\\\Huge A loop")) \%>\%
    ncloop("a","a","angleB=180,loopsize=1,arm=.5,linearc=.2","->")
}
