% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ncput}
\alias{ncput}
\title{Put Label on Line}
\usage{
ncput(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The label to put on the line.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Label on Line
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    cnode(0,4,".5cm","root") \%>\%
    cnode(3,5.5,"4pt","A",star=TRUE) \%>\%
    cnode(3,4,"4pt","B",star=TRUE) \%>\%
    cnode(3,2.5,"4pt","C",star=TRUE) \%>\%
    psset("nodesep=3pt") \%>\%
    ncline("root","A") \%>\%
    naput("above") \%>\%
    ncline("root","B") \%>\%
    ncput("on",star=TRUE) \%>\%
    ncline("root","C") \%>\%
    nbput("below")
}
