% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{newcmykcolor}
\alias{newcmykcolor}
\title{Define New CMYK Color}
\usage{
newcmykcolor(p = NULL, color, num1, num2, num3, num4)
}
\arguments{
\item{p}{The PSTricks object.}

\item{color}{The name of the new color.}

\item{num1, num2, num3, num4}{The cyan-magenta-yellow-black specification (between 0 and 1).}
}
\value{
The updated PSTricks object.
}
\description{
Define New CMYK Color
}
\examples{
newcmykcolor(,"mycolor",0.1,0.2,0.3,0.4)
}
