% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{newgray}
\alias{newgray}
\title{Define New Gray Scale}
\usage{
newgray(p = NULL, color, num)
}
\arguments{
\item{p}{The PSTricks object.}

\item{color}{The name of the new gray scale.}

\item{num}{The scale value (0 is black and 1 is white).}
}
\value{
The updated PSTricks object.
}
\description{
Define New Gray Scale
}
\examples{
newgray(,"gray10",0.1)
}
