% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{newhsbcolor}
\alias{newhsbcolor}
\title{Define New HSB Color}
\usage{
newhsbcolor(p = NULL, color, num1, num2, num3)
}
\arguments{
\item{p}{The PSTricks object.}

\item{color}{The name of the new color.}

\item{num1, num2, num3}{The hue-saturation-brightness specification (between 0 and 1).}
}
\value{
The updated PSTricks object.
}
\description{
Define New HSB Color
}
\examples{
newhsbcolor(,"mycolor",0.1,0.2,0.3)
}
