% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{newrgbcolor}
\alias{newrgbcolor}
\title{Define New RGB Color}
\usage{
newrgbcolor(p = NULL, color, num1, num2, num3)
}
\arguments{
\item{p}{The PSTricks object.}

\item{color}{The name of the new color.}

\item{num1, num2, num3}{The red-green-blue specification (0 is dark and 1 is light).}
}
\value{
The updated PSTricks object.
}
\description{
Define New RGB Color
}
\examples{
newrgbcolor(,"mycolor",0.1,0.2,0.3)
}
