% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{nput}
\alias{nput}
\title{Attach Label to Node}
\usage{
nput(p = NULL, name, stuff, par = NULL, refangle, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{name}{The name of the node.}

\item{stuff}{The label to put on the line.}

\item{par}{PSTricks parameter string.}

\item{refangle}{The reference angle (see \code{\link[=uput]{uput()}}).}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Attach Label to Node
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(4,0,ovalnode(,"B","Node B"),"br") \%>\%
    rput(0,3,rnode(,"A",psframebox(,"Node A")),"tl") \%>\%
    nput("A",paste0(psarcn(,0,0,".4cm",0,-70),
        uput(,0,0,"\\\\texttt{angleA}",-35,labelsep=".4cm")),"labelsep=0",-70) \%>\%
    ncangle("A","B","angleA=-70,angleB=90,armB=1cm,linewidth=1.2pt") \%>\%
    ncput(psframe(,c(0,.35),c(0,.35),"dimen=middle"),"nrot=:U,npos=1")
}
