% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{ovalnode}
\alias{ovalnode}
\title{Put Stuff in an Oval}
\usage{
ovalnode(p = NULL, name, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{name}{The name of the node.}

\item{stuff}{Stuff to put in a box at the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in an Oval
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(8,4,paste(circlenode(,"A","Circle"),"and",ovalnode(,"B","Oval"))) \%>\%
    ncbar("A","B","angle=90")
}
