% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{parabola}
\alias{parabola}
\title{Draw PSTricks Parabola}
\usage{
parabola(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the parabola.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Parabola
}
\examples{
pppicture(PSTricks(),4,3,par="showgrid=true") \%>\%
   parabola(c(1,2), c(1,3), star=TRUE) \%>\%
   psset("xunit=.01") \%>\%
   parabola(c(400,200),c(3,0),arrows="<->")
}
