% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{pcarc}
\alias{pcarc}
\title{Draw an Arc Between Two Nodes}
\usage{
pcarc(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates or names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw an Arc Between Two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    pcarc(c(3,6),c(4,9))
}
