% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{pcbox}
\alias{pcbox}
\title{Enclose Two Nodes in a Box}
\usage{
pcbox(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates or names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Enclose Two Nodes in a Box
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    pcbox(c(3,6),c(4,9))
}
