% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-coil.R
\name{pccoil}
\alias{pccoil}
\title{Draw a Coil between two Nodes}
\usage{
pccoil(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates or names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Coil between two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-coil"),4,3,par="showgrid=true") \%>\%
    pccoil(c(.5,3.5),c(.5,2.5),"coilwidth=.3","<->")
}
