% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-coil.R
\name{pczigzag}
\alias{pczigzag}
\title{Draw a Zigzag between two Nodes}
\usage{
pczigzag(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates or names of the nodes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the zigzag.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Zigzag between two Nodes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-coil"),4,3,par="showgrid=true") \%>\%
    pczigzag(c(.5,3.5),c(.5,2.5),"coilarm=.5,linearc=.1","<->")
}
