% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe PSTricks Object}
\arguments{
\item{lhs, rhs}{A PSTricks object and a function to apply to it.}
}
\description{
Like \code{dplyr}, PSTricks also uses the pipe function, \code{\%>\%}, to pass
information from one function to another. But this is unlike \code{ggplot2}, which
uses the \code{+} operator.
}
\examples{
# Instead of
geom_dots(PSTricks(),aes(x=wt,y=mpg),mtcars)
# one may write
PSTricks() \%>\% geom_dots(aes(x=wt,y=mpg),mtcars)
}
