% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{pnode}
\alias{pnode}
\title{Create Zero-dimensional Node}
\usage{
pnode(p = NULL, x = NULL, y = NULL, name)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the node.}

\item{name}{The name of the node.}
}
\value{
The updated PSTricks object.
}
\description{
Create Zero-dimensional Node
}
\seealso{
\code{\link[=cnode]{cnode()}} for an example.
}
