% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ppappend}
\alias{ppappend}
\title{Append Line to Lines Attribute in the PSTricks Object}
\usage{
ppappend(p, s)
}
\arguments{
\item{p}{The PSTricks object.}

\item{s}{The string to append.}
}
\value{
The updated PSTricks object.
}
\description{
Append Line to Lines Attribute in the PSTricks Object
}
