% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pparg}
\alias{pparg}
\title{Construct \code{pstricks} Argument}
\usage{
pparg(arg = NULL)
}
\arguments{
\item{arg}{Argument.}
}
\value{
Argument string (using curly braces), or empty string if \code{arg} is \code{NULL}.
}
\description{
Construct \code{pstricks} Argument
}
