% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{ppaxis}
\alias{ppaxis}
\title{Draw an X or Y Axis}
\usage{
ppaxis(
  p,
  xory,
  lims,
  label = "label",
  labsep = NULL,
  secondary = FALSE,
  noshow = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{xory}{A character 'x' or 'y' designating which axis to draw.}

\item{lims}{A vector with two elements, the minimum and maximum values for the axis.}

\item{label}{The label to show at the middle of the axis.}

\item{labsep}{The distance between the tickmark labels and the label.}

\item{secondary}{A flag to indicate that a secondary (at the other side) axis should be drawn.}

\item{noshow}{A flag to indicate that values should be scaled with respect to the axis, but that the axis should not be drawn.}
}
\value{
The updated PSTricks object, with attributes \code{xtpos} and \code{ytpos} added for \code{ppgrid()}.
}
\description{
Draw an X or Y Axis
}
\examples{
p <- pppicture(PSTricks(),16,9) \%>\%
    ppticks('x',6,3) \%>\%
    ppticks('y',6,4) \%>\%
    ppaxis('x',c(1,6),"wt") \%>\%
    ppaxis('y',c(10,35),"mpg") ;
    psdots(p,cx(p,mtcars$wt),cy(p,mtcars$mpg))
# Note that p has to have valid axes before using `cx()` or `cy()`
}
