% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ppbuild}
\alias{ppbuild}
\title{Construct \code{pstricks} Macro Command}
\usage{
ppbuild(
  psname,
  x = NULL,
  y = NULL,
  opt = NULL,
  arg = NULL,
  arg1 = NULL,
  arg2 = NULL,
  arg3 = NULL,
  arg4 = NULL,
  arg0 = NULL,
  star = FALSE,
  p = NULL
)
}
\arguments{
\item{psname}{The name of the macro command to construct.}

\item{x, y}{Coordinates.}

\item{opt}{Optional parameters.}

\item{arg, arg1, arg2, arg3, arg4, arg0}{Arguments.}

\item{star}{Flag to indicate starred version.}

\item{p}{The PSTricks object.}
}
\value{
The string or an updated PSTricks object.
}
\description{
Construct \code{pstricks} Macro Command
}
\examples{
ppbuild("ppbuild",1,2,"opt","arg","arg1","arg2","arg3","arg4","arg0",TRUE)
}
