% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{ppclosedoc}
\alias{ppclosedoc}
\title{Close the LaTex Document}
\usage{
ppclosedoc(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Adds a line to the \code{p} object to finish a self-contained LaTeX document.
While this function is exported, it is called automatically when necessary.
}
\examples{
p <- ppclosedoc(ppopendoc(PSTricks()))
}
