% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ppcoords}
\alias{ppcoords}
\title{Construct \code{pstricks} Macro Coordinates}
\usage{
ppcoords(p = NULL, x, y)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates.}
}
\value{
Coordinates string (using parentheses), or empty string if \code{x} or \code{y} is \code{NULL}.
}
\description{
Construct \code{pstricks} Macro Coordinates
}
