% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ppcoords3D}
\alias{ppcoords3D}
\title{Construct \code{pstricks} Macro Coordinates}
\usage{
ppcoords3D(p = NULL, x, y, z)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y, z}{Coordinates.}
}
\value{
Coordinates string (using parentheses), or empty string if \code{x} or \code{y} or \code{z} is \code{NULL}.
}
\description{
Construct \code{pstricks} Macro Coordinates
}
