% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picture.R
\name{ppdefpicture}
\alias{ppdefpicture}
\title{Open a Default Picture}
\usage{
ppdefpicture(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Open a Default Picture
}
\details{
Used by geoms if no picture has been opened.
}
