% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ppgeoms}
\alias{ppgeoms}
\title{Process Geoms}
\usage{
ppgeoms(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Process Geoms
}
\details{
\code{ppgeoms()} is called automatically when the current subplot is closed. The example
given below shows an instance where it is necessary to call it explicitly.
}
\examples{
pppicture(PSTricks(),16,9,data=mtcars) \%>\%
    geom_dots(aes(x=wt,y=mpg),par="linecolor=green") \%>\%
    ppgeoms() \%>\%
    ppsetsecondary('y') \%>\%
    geom_dots(aes(x=wt,y=cyl),par="linecolor=blue")
}
