% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{ppgrid}
\alias{ppgrid}
\title{Draw Grid Lines}
\usage{
ppgrid(p, par = "linestyle=dotted", background = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{par}{PSTricks parameters.}

\item{background}{The optional background color.}
}
\value{
The updated PSTricks object.
}
\description{
Draw Grid Lines
}
\details{
Axes should be drawn before a grid.
Issue: with "linestyle=dotted" multiple dots are drawn at identical locations.
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    newrgbcolor("verylightgray",.9,.9,.9) \%>\%
    ppsetmargins(mrgaxes=0) \%>\%
    ppaxis('x',c(0,1)) \%>\%
    ppaxis('y',c(0,1)) \%>\%
    ppgrid("linestyle=dotted,linecolor=gray",background="verylightgray")
}
