% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{pplegend}
\alias{pplegend}
\title{Add Legend to Plot}
\usage{
pplegend(
  p,
  s,
  par = NULL,
  position = "tr",
  dx = 0,
  dy = 0,
  w = 1,
  labelsep = "10pt"
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{s}{The legend text.}

\item{par}{PSTricks parameter string.}

\item{position}{Position for the legend (may be NULL).}

\item{dx, dy}{x and y offsets w.r.t. default position.}

\item{w}{Width of the \code{psline()} that belongs to the legend text.}

\item{labelsep}{The distance between the line and the label.}
}
\value{
The updated PSTricks object.
}
\description{
Add Legend to Plot
}
\examples{
p <- pppicture(PSTricks(),16,9) \%>\%
    ppaxis('x',c(0,1)) \%>\%
    ppaxis('y',c(0,1)) ;
p <- p \%>\%
    psset("linecolor=green,showpoints=true") \%>\%
    psline(cx(p,seq(0,1,0.2)),cy(p,rep(0.5,5))) \%>\%
    pplegend("top right")
}
