% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pplinewidth}
\alias{pplinewidth}
\title{Set Line Width}
\usage{
pplinewidth(p, linewidth)
}
\arguments{
\item{p}{The PSTricks object.}

\item{linewidth}{The new default line width in mm.}
}
\value{
The updated PSTricks object.
}
\description{
Set Line Width
}
\details{
Parameter linewidth is a special one because it is needed at some
places for proper alignment (\code{geom_frame()},\code{geom_hist()},\code{ppgrid()},\code{pplegend()},\code{cx()},\code{cy()},\code{endP2E()}).
}
\seealso{
\code{\link[=geom_set]{geom_set()}} for an example.
}
