% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{ppmansubplot}
\alias{ppmansubplot}
\title{Set Parameters of Subplot Manually}
\usage{
ppmansubplot(p, x0, y0, hx, hy, ntitle = 1)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x0}{The reference position of the x axis.}

\item{y0}{The reference position of the y axis.}

\item{hx}{The length of the x axis.}

\item{hy}{The length of the y axis.}

\item{ntitle}{Number of lines to reserve for the title.}
}
\value{
The updated PSTricks object.
}
\description{
Set Parameters of Subplot Manually
}
\examples{
pppicture(PSTricks(),20,28,par="showgrid=true") \%>\% ppmansubplot(2,2,8,6) \%>\%
    ppaxis('x',c(0,1)) \%>\% ppaxis('y',c(0,1)) \%>\% pptitle("title")
# note that (x0,y0) is the reference position, not where the axes start
}
\seealso{
\code{\link[=adjx0y0]{adjx0y0()}} to get axis positions.
}
