% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picture.R
\name{ppnewpage}
\alias{ppnewpage}
\title{Close the Current Picture and Open a New One}
\usage{
ppnewpage(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Close the Current Picture and Open a New One
}
\details{
Lower level option values will be reset, but higher level options will not.
}
\examples{
pppicture(PSTricks(engine="latex"),16,9, data=mtcars, par="showgrid=true") \%>\%
    geom_dots(aes(x=wt,y=mpg)) \%>\%
    pptitle("\\\\Large picture 1") \%>\%
    ppnewpage() \%>\%
    geom_dots(aes(x=wt,y=cyl)) \%>\%
    pptitle("\\\\Large picture 2")
# Engine pdflatex gives one page...
}
