% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{ppnewrgbcolor}
\alias{ppnewrgbcolor}
\title{Define New RGB Color(s) from R Color Specification(s)}
\usage{
ppnewrgbcolor(p = NULL, names, values = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{names}{R color names.}

\item{values}{Color values to parse.}
}
\value{
The updated PSTricks object.
}
\description{
Define New RGB Color(s) from R Color Specification(s)
}
\examples{
ppnewrgbcolor(,"blue") # p==NULL works for one color only
}
