% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{ppopendoc}
\alias{ppopendoc}
\title{Open the LaTex Document}
\usage{
ppopendoc(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Adds lines to the \code{p} object to start a self-contained LaTeX document.
While this function is exported, it is called automatically when necessary.
}
\examples{
p <- ppopendoc(PSTricks())
}
