% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ppopt}
\alias{ppopt}
\title{Construct \code{pstricks} Option}
\usage{
ppopt(opt = NULL)
}
\arguments{
\item{opt}{Option.}
}
\value{
Option string (using brackets), or empty string if \code{arg} is \code{NULL}.
}
\description{
Construct \code{pstricks} Option
}
