% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picture.R
\name{pppicture}
\alias{pppicture}
\title{Open a Picture and Prepare for using PSTricks Functions}
\usage{
pppicture(
  p,
  x = NULL,
  y = NULL,
  data = NULL,
  mapping = NULL,
  par = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of upper right corner (and optionally lower left corner).}

\item{data}{Data to use with geoms.}

\item{mapping}{Mapping to use with geoms.}

\item{par}{Parameters for the underlying \code{pspicture} macro (see Voss' latest documentation).}

\item{star}{Flag to indicate that objects should be clipped with respect to the boundaries.}
}
\value{
The updated PSTricks object with initial default values for the attributes
\itemize{
\item datnam - Name of the data for reference.
\item data - Data for geoms.
\item mapping - Mapping for geoms.
\item geoms - List of called geoms.
\item xlim,ylim - Range of \code{x} and \code{y} data.
\item xlab,ylab - Labels for the x and y axes.
\item xlabsep,ylabsep - Distance between tickmark and axes labels.
\item xa,xb,ya,yb - Scaling conversion parameters.
\item xticks - See below.
\item yticks - See below.
\item logx,logy - Flags to indicate logarithmic x and/or y axes.
\item secondx,secondy - Flags to indicate secondary x and/or y axes.
\item pxad,pyad,sxad,syad - Flags to indicate which axes have been drawn.
\item margin - Parameter that determines the layout of a graph.
\item mrgaxes - A factor for the margins between the axes.
\item polar - Flag to indicate whether coordinates should be interpreted as polar.
\item degrees - The number of units in a circle.
\item linewidth - The default line width in cm.
\item picpar - Parameters saved for a possible subsequent \code{pspicture} with \code{ppnewpage()}.
\item psttoeps - Flag to indicate that the PSTtoEPS feature should be used with geoms.
}

\code{xticks} and \code{yticks} are lists with the items
\itemize{
\item nticks -Number of tickmarks; if nticks=0, pretty tickmarks will be determined automatically.
\item mticks - Number of minor tickmarks.
\item nolabels - Flag to indicate that no labels should be printed.
\item extlabs - Flag to indicate that labels at axis extrema should be printed.
\item labels - List of labels instead of numbers to print at the tickmarks.
\item ticklength - The length of the ticks.
\item ticklengthi - The inward length of the ticks (default same as outward).
\item rotation - The rotation for the labels at the tickmarks.
}
}
\description{
Open a Picture and Prepare for using PSTricks Functions
}
\details{
\code{pppicture} is not called \code{pspicture} because of the large difference in functionality.
It is not needed for using PSTricks package per se (as in LaTeX itself).
Most examples use \code{pppicture()}.
}
\seealso{
See \code{\link[=tvput]{tvput()}} for a rare example where \code{pppicture()} is not used.
And see \code{\link[=pspicture]{pspicture()}} for the lower level function.
}
