% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetcartesian}
\alias{ppsetcartesian}
\title{Set Interpretation of Coordinates to Cartesian}
\usage{
ppsetcartesian(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Set Interpretation of Coordinates to Cartesian
}
\seealso{
\code{\link[=ppsetpolar]{ppsetpolar()}}.
}
