% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetlogx}
\alias{ppsetlogx}
\title{Set Flag to use Logarithmic X Axis}
\usage{
ppsetlogx(p, logx = TRUE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{logx}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flag to use Logarithmic X Axis
}
