% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetlogxy}
\alias{ppsetlogxy}
\title{Set Flags to use Logarithmic X and Y Axes}
\usage{
ppsetlogxy(p, logxy = TRUE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{logxy}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flags to use Logarithmic X and Y Axes
}
\seealso{
\code{\link[=geom_abline]{geom_abline()}} for an example.
}
