% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetlogy}
\alias{ppsetlogy}
\title{Set Flag to use Logarithmic Y Axis}
\usage{
ppsetlogy(p, logy = TRUE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{logy}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flag to use Logarithmic Y Axis
}
