% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetmargins}
\alias{ppsetmargins}
\title{Set Overall Margin}
\usage{
ppsetmargins(p, margin = 1, mrgaxes = 1)
}
\arguments{
\item{p}{The PSTricks object.}

\item{margin}{Parameter that determines the layout of a graph.}

\item{mrgaxes}{A factor for the margins between the axes.}
}
\value{
The updated PSTricks object with respect to the attributes \code{margin} and \code{mrgaxes}.
}
\description{
Set Overall Margin
}
\seealso{
\code{\link[=ppgrid]{ppgrid()}} for an example.
}
