% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetnologx}
\alias{ppsetnologx}
\title{Reset Flag to use Logarithmic X Axis}
\usage{
ppsetnologx(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Reset Flag to use Logarithmic X Axis
}
