% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetnologxy}
\alias{ppsetnologxy}
\title{Reset Flags to use Logarithmic X and Y Axes}
\usage{
ppsetnologxy(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Reset Flags to use Logarithmic X and Y Axes
}
