% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetnology}
\alias{ppsetnology}
\title{Reset Flag to use Logarithmic Y Axis}
\usage{
ppsetnology(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Reset Flag to use Logarithmic Y Axis
}
