% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetpolar}
\alias{ppsetpolar}
\title{Set Interpretation of Coordinates to Polar}
\usage{
ppsetpolar(p)
}
\arguments{
\item{p}{The PSTricks object.}
}
\value{
The updated PSTricks object.
}
\description{
Set Interpretation of Coordinates to Polar
}
\seealso{
\code{\link[=degrees]{degrees()}} and \code{\link[=ppsetcartesian]{ppsetcartesian()}}, and \code{\link[=psarcn]{psarcn()}} for an example.
}
