% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetprimary}
\alias{ppsetprimary}
\title{Set Flag to use Primary X or Y Axis}
\usage{
ppsetprimary(p, xory, secondary = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{xory}{A character 'x' or 'y' designating the axis.}

\item{secondary}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flag to use Primary X or Y Axis
}
