% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetprimaryy}
\alias{ppsetprimaryy}
\title{Set Flag to use Primary Y Axis}
\usage{
ppsetprimaryy(p, secondary = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{secondary}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flag to use Primary Y Axis
}
