% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetpsttoeps}
\alias{ppsetpsttoeps}
\title{Set Flag to use PSTtoEPS Feature}
\usage{
ppsetpsttoeps(p, psttoeps = TRUE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{psttoeps}{A flag to indicate that the PSTtoEPS feature should be used with geoms.}
}
\description{
Set Flag to use PSTtoEPS Feature
}
\details{
The PSTtoEPS feature is explained in the original manual in section 55.
It may be used for efficient EPS file processing, in particular in cases
where TeX's capacity becomes exceeded with many plotting commands.
It is needed only for the "latex" engine; "xelatex" and "lualatex" do not
handle it properly.
The "pstpkgs="pst-eps" must be used when creating the \code{PSTricks()} object.
}
