% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetsecondary}
\alias{ppsetsecondary}
\title{Set Flag to use Secondary X or Y Axis}
\usage{
ppsetsecondary(p, xory, secondary = TRUE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{xory}{A character 'x' or 'y' designating the axis.}

\item{secondary}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flag to use Secondary X or Y Axis
}
\seealso{
\code{\link[=ppgeoms]{ppgeoms()}} for an example.
}
