% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetsecondaryx}
\alias{ppsetsecondaryx}
\title{Set Flag to use Secondary X Axis}
\usage{
ppsetsecondaryx(p, secondary = TRUE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{secondary}{The flag.}
}
\value{
The updated PSTricks object.
}
\description{
Set Flag to use Secondary X Axis
}
\seealso{
\code{\link[=ppgeoms]{ppgeoms()}} for an example.
}
