% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetxlabsep}
\alias{ppsetxlabsep}
\alias{ppxlabsep}
\title{Set x label separation distance}
\usage{
ppsetxlabsep(p, labsep = 0.7)

ppxlabsep(p, labsep = 0.7)
}
\arguments{
\item{p}{The PSTricks object.}

\item{labsep}{The distance.}
}
\value{
The updated PSTricks object.
}
\description{
Set x label separation distance
}
\examples{
geom_line(PSTricks(),data=data.frame(x=c(4,0,2),y=c(2,1,0)),
    par="linewidth=2pt,linearc=.25,arrows=->") \%>\%
    ppsetxlabsep(1.5) \%>\% ppsetylabsep(2)
}
\seealso{
\code{\link[=geom_line]{geom_line()}} to view the default distances.
}
