% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ppsetylabsep}
\alias{ppsetylabsep}
\alias{ppylabsep}
\title{Set y label separation distance}
\usage{
ppsetylabsep(p, labsep = 1)

ppylabsep(p, labsep = 1)
}
\arguments{
\item{p}{The PSTricks object.}

\item{labsep}{The distance.}
}
\value{
The updated PSTricks object.
}
\description{
Set y label separation distance
}
\seealso{
\code{\link[=ppsetxlabsep]{ppsetxlabsep()}} for an example.
}
