% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{pptitle}
\alias{pptitle}
\title{Set Plot Title}
\usage{
pptitle(p, title, dx = 0, dy = 0)
}
\arguments{
\item{p}{The PSTricks object.}

\item{title}{The title.}

\item{dx, dy}{Offset with respect to the default position (top left).}
}
\value{
The updated PSTricks object.
}
\description{
Set Plot Title
}
\details{
The title is shown using \code{uput()}.
}
\seealso{
\code{\link[=ppmansubplot]{ppmansubplot()}} for an example.
}
