% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ppwrite}
\alias{ppwrite}
\title{Write Assembled PSTricks Picture(s) to a File}
\usage{
ppwrite(
  p,
  filename = NULL,
  topdf = TRUE,
  crop = FALSE,
  topng = FALSE,
  dsf = 4,
  toeps = FALSE,
  clean = TRUE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{filename}{The name of the .tex file to write the document to (by default the name of the script, or "pp" when interactive).}

\item{topdf}{Flag to specify if a .pdf should be generated by the engine as specified with \code{PSTricks()}.}

\item{crop}{Flag if a cropped version with name \code{-crop.pdf} should be created.}

\item{topng}{Flag to specify if the .pdf should be converted to a .png.}

\item{dsf}{DownScaleFactor for Ghostscript when converting to .png (resolution is 4x72=288 pixels per inch).}

\item{toeps}{Flag to specify if an .eps should be generated (using latex and dvips -E).}

\item{clean}{Flag to specify if intermediate files should be deleted after generating the .pdf.}
}
\value{
Nothing.
}
\description{
\code{ppwrite()} is used to write the assembled LaTeX document to a file.
It does not return the PSTricks object, as it will no longer be useful
(a new \code{PSTricks()} call is needed).
\code{ppwrite} may be called automatically by R via \code{print} (\code{print.PSTricks}).
}
\examples{
ppwrite(pppicture(PSTricks(engine="pdflatex"),par="showgrid=true"))
# where the "pdflatex" engine is the only one showing the grid labels
# with a full A4 picture.
}
