% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks.R
\name{ppyticks}
\alias{ppyticks}
\alias{yticks}
\title{Define Major and Minor Tickmarks at the Y Axis}
\usage{
ppyticks(
  p,
  nticks = 0,
  mticks = 0,
  nolabels = FALSE,
  extlabs = FALSE,
  labels = NULL,
  rotation = 0,
  ticklength = 0.2,
  ticklengthi = NULL
)

yticks(
  p,
  nticks = 0,
  mticks = 0,
  nolabels = FALSE,
  extlabs = FALSE,
  labels = NULL,
  rotation = 0,
  ticklength = 0.2,
  ticklengthi = NULL
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{nticks}{Number of tickmarks; if nticks=0, pretty tickmarks will be determined automatically.}

\item{mticks}{Number of minor tickmarks.}

\item{nolabels}{Flag to indicate that no labels should be printed.}

\item{extlabs}{Flag to indicate that labels at axis extrema should be printed (however labels cannot be used).}

\item{labels}{List of labels instead of numbers to print at the tickmarks.}

\item{rotation}{The rotation for the labels at the tickmarks.}

\item{ticklength}{The length of the ticks.}

\item{ticklengthi}{\itemize{
\item The inward length of the ticks (default same as outward).
}}
}
\value{
The updated PSTricks object.
}
\description{
Define Major and Minor Tickmarks at the Y Axis
}
\examples{
PSTricks() \%>\%
    geom_dots(aes(x=wt,y=mpg),mtcars) \%>\%
    ylim(10,35) \%>\%
    yticks(6,0)
}
