% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{print.PSTricks}
\alias{print.PSTricks}
\title{\code{print} a PSTricks Object}
\usage{
\method{print}{PSTricks}(x, ...)
}
\arguments{
\item{x}{The PSTricks object.}

\item{...}{Parameters for \code{ppwrite}.}
}
\description{
\code{print} a PSTricks Object
}
