% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psarc}
\alias{psarc}
\title{Draw PSTricks Arc}
\usage{
psarc(
  p = NULL,
  x = NULL,
  y = NULL,
  radius,
  angleA,
  angleB,
  par = NULL,
  arrows = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the arc.}

\item{radius}{Radius of the arc.}

\item{angleA, angleB}{Start and end angles of the arc.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Arc
}
\examples{
pppicture(PSTricks(),3,2,par="showgrid=true") \%>\%
    psarc(1.5,1.5,1.5,215,0,"showpoints=true",star=TRUE)
}
