% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psarcn}
\alias{psarcn}
\title{Draw PSTricks Arc Clockwise}
\usage{
psarcn(
  p = NULL,
  x = NULL,
  y = NULL,
  radius,
  angleA,
  angleB,
  par = NULL,
  arrows = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the arc.}

\item{radius}{Radius of the arc.}

\item{angleA, angleB}{End and start angles of the arc.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Arc Clockwise
}
\examples{
pppicture(PSTricks(),4,3,par="showgrid=true") \%>\%
    ppsetpolar() \%>\%
    psline(c(4,0,4),c(50,0,10),"linewidth=2pt") \%>\%
    psarcn(0,0,3,50,10,"arcsepB=2pt",arrows="<-")
}
