% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-plot.R
\name{psaxes}
\alias{psaxes}
\title{Draw PSTricks Axes}
\usage{
psaxes(p = NULL, x, y, par = NULL, arrows = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the axes.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Axes
}
\examples{
pppicture(PSTricks(pstpkgs="pst-plot"),4,3,par="showgrid=true") \%>\%
    psaxes(c(2,0,4), c(1,0,3),
        "linewidth=1.2pt,labels=none,ticks=none", "<->")
# observe interesting showgrid
}
