% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psbezier}
\alias{psbezier}
\title{Draw PSTricks Bezier Curve}
\usage{
psbezier(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the line segment(s).}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Bezier Curve
}
\examples{
pppicture(PSTricks(),4,4) \%>\%
    psbezier(c(0,1,2,4),c(0,4,1,3.5),"linewidth=2pt,showpoints=true","->")
}
