% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psccurve}
\alias{psccurve}
\title{Draw PSTricks Closed Curve}
\usage{
psccurve(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the curve.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Closed Curve
}
\examples{
pppicture(PSTricks(),4,1,par="showgrid=true") \%>\%
    psccurve(c(.5,3.5,3.5,.5),c(0,1,0,1),"showpoints=true")
}
