% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pscirclebox}
\alias{pscirclebox}
\title{Draw a Circle Box}
\usage{
pscirclebox(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The stuff to put in the box.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw a Circle Box
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,pscirclebox(,"{\\\\begin{tabular}{c} You are \\\\\\\\ here \\\\end{tabular}}"))
}
