% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pscircleOA}
\alias{pscircleOA}
\title{Draw PSTricks Circle}
\usage{
pscircleOA(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the center of the circle and one point on the circle.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Circle
}
\examples{
pppicture(PSTricks(engine="latex"),8,8,par="showgrid=true") \%>\%
    pscircleOA(c(6,4),c(4,4)) \%>\%
    pscircleOA(c(4,4),c(6,4),"linecolor=blue") \%>\%
    pscircleOA(c(3,4),c(5,4),"linewidth=2pt,linecolor=yellow") \%>\%
    pscircleOA(c(2,4),c(4,4),"opacity=0.3,linecolor=red",TRUE)
}
