% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-coil.R
\name{pscoil}
\alias{pscoil}
\title{Draw PSTricks Coil}
\usage{
pscoil(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the coil.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the coil.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Coil
}
\examples{
pppicture(PSTricks(pstpkgs="pst-coil"),c(-1,5),c(-1,3),par="showgrid=true") \%>\%
    pscoil(4,2,"coilarm=.5cm,linewidth=1.5pt,coilwidth=.5cm","<-|")
}
