% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pscurve}
\alias{pscurve}
\title{Draw PSTricks Curve}
\usage{
pscurve(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the curve.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Curve
}
\examples{
pppicture(PSTricks(),4,2,par="showgrid=true") \%>\%
    pscurve(c(0,0.7,3.3,4,0.4),c(1.3,1.8,0.5,1.6,0.4),
        "showpoints=true","<->")
}
