% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pscustom}
\alias{pscustom}
\title{Custom graphics}
\usage{
pscustom(p = NULL, commands, par = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{commands}{Commands to call.}

\item{par}{PSTricks parameter string.}
}
\value{
The updated PSTricks object.
}
\description{
Custom graphics
}
\examples{
PSTricks() \%>\%
    pppicture(4,3,par="showgrid=true") \%>\%
    pscustom(paste0(pscurve(,c(0,1,2,4),c(2,2.5,1.5,3)),
                    pscurve(,c(4,3,2,1,0),c(1,0.5,1,0,0.5),"liftpen=1")),
             "linewidth=2pt,fillstyle=solid,fillcolor=gray")
}
