% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psdiamond}
\alias{psdiamond}
\title{Draw PSTricks Diamond}
\usage{
psdiamond(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the diamond.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Diamond
}
\examples{
pppicture(PSTricks(),4,2,par="showgrid=true") \%>\%
    psdiamond(c(2,1.5),c(1,1),"framearc=.3,fillstyle=solid,fillcolor=lightgray")
}
