% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psdot}
\alias{psdot}
\title{Draw PSTricks Dot}
\usage{
psdot(p = NULL, x = NULL, y = NULL, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the dot.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Dot
}
\examples{
pppicture(PSTricks(),2,2) \%>\%
    psdot(1,1)
}
