% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psdots}
\alias{psdots}
\title{Draw PSTricks Dots}
\usage{
psdots(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the dots.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Dots
}
\examples{
pppicture(PSTricks(),2,2) \%>\%
    psdots(c(0,1,2),c(1,1,1),"dotstyle=Bo")
}
