% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psecurve}
\alias{psecurve}
\title{Draw PSTricks Extended Curve}
\usage{
psecurve(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the curve.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Extended Curve
}
\examples{
pppicture(PSTricks(),4,4,par="showgrid=true") \%>\%
    psecurve(c(.125,.25,.5,1,2,4,8),c(8,4,2,1,.5,.25,.125),
        "showpoints=true")
}
