% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psellipse}
\alias{psellipse}
\title{Draw PSTricks Ellipse}
\usage{
psellipse(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the center of and the horizontal and vertical radii.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Ellipse
}
\examples{
pppicture(PSTricks(),c(-1,2),c(-1,1),par="showgrid=true") \%>\%
    psellipse(c(.5,1.5),c(0,1),"fillcolor=lightgray")
}
