% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psellipticarc}
\alias{psellipticarc}
\title{Draw PSTricks Elliptic Arc}
\usage{
psellipticarc(
  p = NULL,
  x,
  y,
  angleA,
  angleB,
  par = NULL,
  arrows = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the elliptic arc.}

\item{angleA, angleB}{Start and end angles of the arc.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Elliptic Arc
}
\examples{
pppicture(PSTricks(),c(-1,2),c(-1,1),par="showgrid=true") \%>\%
    psellipticarc(c(.5,1.5),c(0,1),215,0,"showpoints=true,arrowscale=2","->")
}
