% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psframe}
\alias{psframe}
\title{Draw PSTricks Frame}
\usage{
psframe(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the frame.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Frame
}
\examples{
pppicture(PSTricks(),4,2,par="showgrid=true") \%>\%
    psframe(4,2,"linewidth=2pt,framearc=.3,fillstyle=solid,fillcolor=lightgray") \%>\%
    psframe(c(1,2),c(.5,1.5),"linecolor=white",star=TRUE)
}
