% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psframebox}
\alias{psframebox}
\title{Put Stuff in a Box with a Frame}
\usage{
psframebox(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The stuff to put in the box.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Box with a Frame
}
\examples{
pppicture(PSTricks(),3,2) \%>\%
    pspolygon(c(0,3,3,2),c(0,0,2,2),"fillcolor=gray,fillstyle=crosshatch*") \%>\%
    rput(2,1,psframebox(,"Label","framearc=.3",star=TRUE))
}
