% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psgrid}
\alias{psgrid}
\title{Draw PSTricks Grid}
\usage{
psgrid(p = NULL, x = NULL, y = NULL, par = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the grid.}

\item{par}{PSTricks parameter string.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Grid
}
\examples{
PSTricks() \%>\%
    pppicture(c(-2,4),c(-2,3)) \%>\%
    psgrid(c(0,-1,3), c(0,-1,2)) \%>\%
    pppicture(c(-1,3),c(-1,2)) \%>\%
    psgrid()
}
