% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psline}
\alias{psline}
\title{Draw PSTricks Line}
\usage{
psline(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the line segment(s).}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the line.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Line
}
\examples{
pppicture(PSTricks(),4,2,par="showgrid=true") \%>\%
    psline(c(4,0,2),c(2,1,0),"linewidth=2pt,linearc=.25","->")
}
\seealso{
\code{\link[=geom_line]{geom_line()}} for the version with scaling.
}
