% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psovalbox}
\alias{psovalbox}
\title{Put Stuff in an Oval Box}
\usage{
psovalbox(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The stuff to put in the box.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in an Oval Box
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,paste0("\\\\parbox{3.75cm}{At the introductory price of ",
        psovalbox(,"\\\\$13.99","boxsep=false,linecolor=darkgray"),
        ", it pays to act now!}"))
}
