% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pspicture}
\alias{pspicture}
\title{Begin Picture Environment}
\usage{
pspicture(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of upper right corner (and optionally lower left corner).}

\item{par}{Parameters (see Voss' latest documentation).}

\item{star}{Flag to indicate that objects should be clipped with respect to the boundaries.}
}
\value{
The updated PSTricks object.
}
\description{
Begin Picture Environment
}
\details{
Available, but see \code{pppicture()}.
}
\examples{
PSTricks(engine="pdflatex") \%>\%
    psset("linecolor=red") \%>\%
    pspicture(1,1,"showgrid") \%>\%
    rput(0,0,
         paste(pspicture(,1,1,star=TRUE),
               psline(,c(-1,2),c(-1,2)),
               endpspicture()),
         "lb") \%>\%
    endpspicture()
# Example found on the internet for clipping while showing labels
}
