% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pspolygon}
\alias{pspolygon}
\title{Draw PSTricks Polygon}
\usage{
pspolygon(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the line segment(s).}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Polygon
}
\examples{
pppicture(PSTricks(),4,2,par="showgrid=true") \%>\%
    pspolygon(c(0,1),c(2,2),"linewidth=1.5pt") \%>\%
    pspolygon(c(1,1,4,4),c(0,2,0,2),"linearc=.2", star=TRUE)
}
