% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psscalebox}
\alias{psscalebox}
\title{Scale Box}
\usage{
psscalebox(p = NULL, stuff, num1, num2)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{Stuff to scale.}

\item{num1, num2}{Numbers to scale horizontally and vertically}
}
\value{
The updated PSTricks object.
}
\description{
Scale Box
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,psscalebox(,"Big and long",4,2))
}
