% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psscaleboxto}
\alias{psscaleboxto}
\title{Scale Box To}
\usage{
psscaleboxto(p = NULL, x, y, stuff)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Width and height to scale to.}

\item{stuff}{Stuff to rotate.}
}
\value{
The updated PSTricks object.
}
\description{
Scale Box To
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,psscaleboxto(,4,2,"Big and long"))
}
