% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psset}
\alias{psset}
\title{Set Any Native PSTricks Option}
\usage{
psset(p = NULL, s)
}
\arguments{
\item{p}{The PSTricks object.}

\item{s}{A string with \code{par=value} specifications (comma separated).}
}
\value{
The updated PSTricks object.
}
\description{
Set Any Native PSTricks Option
}
\examples{
psset(,"linewidth=0.1mm")
}
