% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psshadowbox}
\alias{psshadowbox}
\title{Put Stuff in a Box with a Frame and a Shadow}
\usage{
psshadowbox(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The stuff to put in the box.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Box with a Frame and a Shadow
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,"\\\\psshadowbox{\\\\textbf{Great Idea!!}}")
}
