% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{psTextFrame}
\alias{psTextFrame}
\title{Draw PSTricks Text Frame}
\usage{
psTextFrame(p = NULL, x, y, text, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the frame.}

\item{text}{Text to display in the frame.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Text Frame
}
\examples{
pppicture(PSTricks(),8,6,par="showgrid=true") \%>\%
    psTextFrame(c(0,4),c(0.5,1.5),"Hallo","linecolor=lightgray,ref=l")
}
