% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pstriangle}
\alias{pstriangle}
\title{Draw PSTricks Triangle}
\usage{
pstriangle(p = NULL, x, y, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the triangle.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Triangle
}
\examples{
pppicture(PSTricks(),4,2,par="showgrid=true") \%>\%
    pstriangle(c(2,4),c(.5,1),"gangle=10", star=TRUE)
}
