% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{PSTricks}
\alias{PSTricks}
\title{Create a PSTricks Object}
\usage{
PSTricks(
  x = NULL,
  y = NULL,
  engine = c("default", "lualatex", "xelatex", "pdflatex", "latex"),
  paper = c("default", "a4", "letter"),
  landscape = FALSE,
  center = TRUE,
  packages = NULL,
  pstpkgs = NULL,
  familydefault = NULL,
  tmpdir = "."
)
}
\arguments{
\item{x}{Width of paper (default A4).}

\item{y}{Height of paper (default A4).}

\item{engine}{Engine to produce a .pdf from the output .tex file. One of "lualatex" (default), "xelatex", "pdflatex", and "latex". No pdf will be produced if the engine name is not recognized.}

\item{paper}{Paper size specification. One of "a4" (default) or "letter".}

\item{landscape}{Flag to indicate landscape paper.}

\item{center}{Flag to use LaTeX offsets to center pictures based on the first one.}

\item{packages}{Font or other packages to load (default default).}

\item{pstpkgs}{PSTricks packages in addition to pstricks itself (default none).}

\item{familydefault}{Familydefault (default \\sfdefault).}

\item{tmpdir}{Temporary directory for the PSTtoEPS feature.}
}
\value{
An initial PSTricks object with attributes
\itemize{
\item docOpened - A flag indicating that the LaTex document has been opened (in the \code{lines} attribute`).
\item picOpened - A flag indicating that the \code{pspicture} PSTricks environment has been opened.
\item paperx - The horizontal paper size in cm.
\item papery - The vertical paper size in cm.
\item x - The horizontal picture size in cm.
\item y - The vertical picture size in cm.
\item landscape - A flag indicating portrait or landscape output mode.
\item center - A flag indicating that the \code{pspicture} will be centered on the paper.
\item config - A list of configuration items (see below).
\item lines - A list of LaTeX lines to be created.
}

The configuration list may consist of the following items:
\itemize{
\item engine - The engine used to process the generated .tex file.
\item familydefault - The default font family.
\item packages - A list of additional LaTeX packages to be used.
\item paper - The type of paper, for example "a4" or "letter".
\item pstpkgs - A list of additional PSTricks packages (normally only "pstricks.sty").
\item tmpdir - The temporary directory for the \code{PSTtoEPS} feature.
}
}
\description{
Create a PSTricks Object
}
\examples{
names(PSTricks())
}
