% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{pswedge}
\alias{pswedge}
\title{Draw PSTricks Wedge}
\usage{
pswedge(
  p = NULL,
  x = NULL,
  y = NULL,
  radius,
  angle1,
  angle2,
  par,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the center of and the horizontal and vertical radii.}

\item{radius}{Radius of the wedge.}

\item{angle1, angle2}{End and start angles of the wedge.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Wedge
}
\examples{
pppicture(PSTricks(),2,2,par="showgrid=true") \%>\%
    pswedge(0,0,2,0,70,"linecolor=gray,linewidth=2pt,fillstyle=solid")
}
