% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-coil.R
\name{pszigzag}
\alias{pszigzag}
\title{Draw PSTricks Zigzag}
\usage{
pszigzag(p = NULL, x, y, par = NULL, arrows = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the zigzag.}

\item{par}{PSTricks parameter string.}

\item{arrows}{Arrows at the end of the zigzag.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Zigzag
}
\examples{
pppicture(PSTricks(pstpkgs="pst-coil"),c(-1,5),c(-1,1),par="showgrid=true") \%>\%
    pszigzag(4,0,"coilarm=.5,linearc=.1","<->")
# Note that the zigzag is drawn partly outside the pppicture.
}
