% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{qdisk}
\alias{qdisk}
\title{Draw PSTricks Disk}
\usage{
qdisk(p = NULL, x, y, radius)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the center of the disk.}

\item{radius}{Radius of the disk.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Disk
}
\examples{
pppicture(PSTricks(),4,6) \%>\%
    psset("linecolor=gray") \%>\%
    qdisk(2,3, 4*2.54/72)
}
