% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{qline}
\alias{qline}
\title{Draw PSTricks Line Segment}
\usage{
qline(p = NULL, x, y)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the line segment.}
}
\value{
The updated PSTricks object.
}
\description{
Draw PSTricks Line Segment
}
\examples{
pppicture(PSTricks(),2,1,par="showgrid=true") \%>\%
    qline(c(0,2),c(0,1))
}
