% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{rnode}
\alias{rnode}
\title{Put Stuff in a Box at a Node}
\usage{
rnode(p = NULL, name, stuff, refpoint = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{name}{The name of the node.}

\item{stuff}{Stuff to put in a box at the node.}

\item{refpoint}{The reference point (see \code{\link[=rput]{rput()}}).}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Box at a Node
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(8,4,paste("\\\\Large",rnode(,"A","sp"),"\\\\hskip 2cm",rnode(,"B","Bit"))) \%>\%
    ncline("A","B")
}
