% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{rotatedown}
\alias{rotatedown}
\title{Rotate Box Down}
\usage{
rotatedown(p = NULL, stuff)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{Stuff to rotate.}
}
\value{
The updated PSTricks object.
}
\description{
Rotate Box Down
}
\seealso{
\code{\link[=rotateleft]{rotateleft()}} for an example.
}
