% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{rotateleft}
\alias{rotateleft}
\title{Rotate Box Left}
\usage{
rotateleft(p = NULL, stuff)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{Stuff to rotate.}
}
\value{
The updated PSTricks object.
}
\description{
Rotate Box Left
}
\examples{
pppicture(PSTricks(),16,9) \%>\%
    rput(8,4,paste("\\\\Large\\\\bfseries",
        rotateleft(,"Left"),rotatedown(,"Down"),rotateright(,"Right")))
}
