% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{rotateright}
\alias{rotateright}
\title{Rotate Box Right}
\usage{
rotateright(p = NULL, stuff)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{Stuff to rotate.}
}
\value{
The updated PSTricks object.
}
\description{
Rotate Box Right
}
\seealso{
\code{\link[=rotateleft]{rotateleft()}} for an example.
}
