% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{rput}
\alias{rput}
\title{Put Stuff at Refpoint}
\usage{
rput(
  p = NULL,
  x = NULL,
  y = NULL,
  stuff,
  refpoint = NULL,
  rotation = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the stuff (may be omitted if \code{rotation} is present).}

\item{stuff}{Stuff to put at the reference point.}

\item{refpoint}{The reference point for the stuff.}

\item{rotation}{Rotation to apply to the stuff.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff at Refpoint
}
\examples{
pppicture(PSTricks(),c(-1,4),c(-1,4)) \%>\%
    rput(stuff=paste0(psframe(,c(-1,2),c(0,1)),
        rput(,2,1,"\\\\emph{stuff}","br","*0")),rotation=34)
}
