% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sifelse}
\alias{sifelse}
\title{Conditional Object Selection}
\usage{
sifelse(test, rt, rf)
}
\arguments{
\item{test}{An object which can be coerced to logical mode.}

\item{rt}{Return value if \code{test} is true.}

\item{rf}{Return value if \code{test} is false.}
}
\value{
Appropriate return value.
}
\description{
Conditional Object Selection
}
\details{
This is like \code{ifelse}, but for a scalar test, and any object may be returned.
}
