% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2e.R
\name{startP2E}
\alias{startP2E}
\title{Start PSTtoEPS Feature}
\usage{
startP2E(p, fileplot = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{fileplot}{Flag to indicate \code{cat}ed values will be used for \code{fileplot}.}
}
\value{
The updated PSTricks object.
}
\description{
Start PSTtoEPS Feature
}
