% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{tbput}
\alias{tbput}
\title{Put Stuff on Line}
\usage{
tbput(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The label to put on the line.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff on Line
}
\seealso{
\code{\link[=tlput]{tlput()}} for an example.
}
