% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{thput}
\alias{thput}
\title{Put Stuff on Line}
\usage{
thput(p = NULL, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{stuff}{The label to put on the line.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff on Line
}
\examples{
PSTricks(engine="lualatex",pstpkgs="pst-node") \%>\%
    ppappend("\\\\[") \%>\%
    ppappend("\\\\setlength{\\\\arraycolsep}{1.1cm}") \%>\%
    ppappend("\\\\begin{array}{cc}") \%>\%
    ppappend(paste(Rnode(,"a","(X-A)"),"&",Rnode(,"b","A"),"\\\\\\\\[1.5cm]")) \%>\%
    ppappend(paste(Rnode(,"c","x"),"&",Rnode(,"d","\\\\tilde{X}"))) \%>\%
    ppappend("\\\\end{array}") \%>\%
    psset("nodesep=5pt,arrows=->") \%>\%
    everypsbox("\\\\scriptstyle") \%>\%
    ncline("a","c") \%>\% thput("h") \%>\%
    ncline("a","b") \%>\% thput("h") \%>\%
    ncline("b","d") \%>\% tvput("v") \%>\%
    ncline("c","d") \%>\% tvput("v") \%>\%
    ppappend("\\\\]")
}
