% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst-node.R
\name{trinode}
\alias{trinode}
\title{Put Stuff in a Triangle}
\usage{
trinode(p = NULL, name, stuff, par = NULL, star = FALSE)
}
\arguments{
\item{p}{The PSTricks object.}

\item{name}{The name of the node.}

\item{stuff}{Stuff to put in a box at the node.}

\item{par}{PSTricks parameter string.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff in a Triangle
}
\examples{
pppicture(PSTricks(pstpkgs="pst-node"),c(-2,14),c(-2,10),par="showgrid=true") \%>\%
    rput(0,3,dianode(,"A","Diamond"),"tl") \%>\%
    rput(4,0,trinode(,"B","Triangle","trimode=L"),"br") \%>\%
    nccurve("A","B","angleA=-135,angleB=90")
}
