% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstricks.R
\name{uput}
\alias{uput}
\title{Put Stuff as Label}
\usage{
uput(
  p = NULL,
  x = NULL,
  y = NULL,
  stuff,
  refangle = NULL,
  rotation = NULL,
  labelsep = NULL,
  star = FALSE
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{x, y}{Coordinates of the stuff (may be omitted if \code{rotation} is present).}

\item{stuff}{Stuff to put at the reference point.}

\item{refangle}{The reference angle.}

\item{rotation}{Rotation to apply to the stuff.}

\item{labelsep}{Distance between coordinates and the stuff.}

\item{star}{Flag to indicate starred version.}
}
\value{
The updated PSTricks object.
}
\description{
Put Stuff as Label
}
\examples{
pppicture(PSTricks(),3,3) \%>\%
    qdisk(1,1,"1pt") \%>\%
    uput(1,1,"(1,1)",45)
}
