% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{xlab}
\alias{xlab}
\title{Set x Axis Label}
\usage{
xlab(p, lab)
}
\arguments{
\item{p}{The PSTricks object.}

\item{lab}{x axis label.}
}
\value{
The updated PSTricks object.
}
\description{
Set x Axis Label
}
