% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{xlim}
\alias{xlim}
\title{Set x Axis Limits}
\usage{
xlim(p, xl = NULL, xu = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{xl, xu}{Lower and upper axis limits.}
}
\value{
The updated PSTricks object.
}
\description{
Set x Axis Limits
}
\seealso{
See \code{\link[=geom_curve]{geom_curve()}} for an example.
}
