% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{xyaspect}
\alias{xyaspect}
\title{Calculate x,y for pppicture given x,y (in p) to get hy = aspect*hx}
\usage{
xyaspect(
  p,
  aspect = 1,
  nx = 1,
  ny = 1,
  nxaxes = 1,
  nyaxes = 1,
  ntitle = 1,
  width = 1,
  height = 1,
  margin = 1
)
}
\arguments{
\item{p}{The PSTricks object.}

\item{aspect}{Desired aspect ratio of axes.}

\item{nx}{Number of plots in the x direction (if NULL, increment n automatically).}

\item{ny}{Number of plots in the y direction.}

\item{nxaxes}{Number of x axes to make space for.}

\item{nyaxes}{Number of y axes to make space for.}

\item{ntitle}{Number of title lines to make space for.}

\item{width}{Number of subplots to occupy in the x direction.}

\item{height}{Number of subplots to occupy in the y direction.}

\item{margin}{Margin.}
}
\value{
The updated PSTricks object.
}
\description{
Calculate x,y for pppicture given x,y (in p) to get hy = aspect*hx
}
\examples{
PSTricks() \%>\% xyaspect(ntitle=0) \%>\% pppicture(par="showgrid=true") \%>\%
    geom_dots(aes(x=wt,y=mpg),mtcars)
}
