% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ylab}
\alias{ylab}
\title{Set y Axis Label}
\usage{
ylab(p, lab)
}
\arguments{
\item{p}{The PSTricks object.}

\item{lab}{y axis label.}
}
\value{
The updated PSTricks object.
}
\description{
Set y Axis Label
}
