% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ylim}
\alias{ylim}
\title{Set y Axis Limits}
\usage{
ylim(p, yl = NULL, yu = NULL)
}
\arguments{
\item{p}{The PSTricks object.}

\item{yl, yu}{Lower and upper axis limits.}
}
\value{
The updated PSTricks object.
}
\description{
Set y Axis Limits
}
\seealso{
See \code{\link[=geom_curve]{geom_curve()}} for an example.
}
